Added particle id to facilitate identification of particles.
Fixed a possible bug in cell. The cell size was not getting updated
based on the smoothing length at all. A default value of 0.1 was being
used for all runs.

Added log messages to integrator and sph_calc to identify neighbors
for a given particle. This would come in handy for comparison of a
serial run with a parallel run.